Add support for inferring & completing main class
Inferring the class means that when the mainClass is omitted, we check
the JAR manifest of the given G:A:V for a Main-Class entry and use that.

Completing the class means that you can now pass a partial string and
jrun will scan all JAR contents for a matching fragment, and use that.

This change now makes it possible to pass only G:A for the endpoint.
We also support passing G:A:V. We differentiate from the G:A:M triple
by heuristic: if the third token starts with a number, or is RELEASE or
LATEST, then we assume it is a version; otherwise it must be a class.