Make the autocomplete logic less automagical
Check out this example:

    $ jrun org.codehaus.groovy:groovy-all:org.codehaus.groovy.tools.shell.Main
    Autocompleted main class: org.codehaus.groovy.tools.shell.Main$_main_closure1
    Error: Main method not found in class org.codehaus.groovy.tools.shell.Main$_main_closure1, please define the main method as:
       public static void main(String[] args)
    or a JavaFX application class must extend javafx.application.Application

Of course, the class I wanted was the very class I stated,
but the autocompletion logic thought it knew better.

To avoid this issue, we now require an @ prefix before class fragments
intended to be autocompletion. This will also improve startup
performance in the case where the fully qualified class name is given.