Harvest settings from .jrunrc file
In particular, we glean the following new settings:

* cacheDir, defaulting to ~/.jrun
* m2Repo, defaulting to ~/.m2/repository

There is almost certainly a way to avoid requiring the user to define
m2Repo (Maven knows where to look for it; we want to look there too),
but I'll figure it out later.

As for cacheDir, this is the handy place where we stash the workspaces.
At least: it will be, after the next commit.