Make jrun work with both Cygwin and msys2
Ironically, using "cmd /c cd" to make things more friendly toward Git
for Windows actually did the opposite: it caused git bash to hang.
It turns out that msys2 ships cygpath anyway, so we can just use it.

There is one wrinkle: cygpath does not do the right thing if the given
path ends in an asterisk. So we exclude the filename here, getting only
the path for the directory in question, and then tack on the filename.

Fixes #7.