Add --resolve-only option
When the option is provided, only the workspace is created and the location is printed to the command line. This can then be used, for example, as classpath for another tool.

```shell
$ jgo --resolve-only org.scijava:parsington
/home/hanslovskyp/.jgo/org/scijava/parsington/RELEASE
$ ls /home/hanslovskyp/.jgo/org/scijava/parsington/RELEASE
mainClass  parsington-1.0.4.jar  pom.xml
```

This is experimental (for now) but might be merged at some point in the future