get_coordinates: fix concatenation syntax
Having "+ xxx" on a new line looked like valid syntax at first glance
but of course is not. The concatenation was just resolving as
independent lines of code and never actually adding versions or
classifiers to the jgo coordinates.

This resulted in all lists of the same components hashing to the same
string, regardless of those component versions.

Closes #68