Remove obsolete release script
It doesn't work with the current project structure (we switched from
setup.py to pyproject.toml, for example), and I don't want to
update it. For me, it is easier to just do the release by hand:

1. Remove the .dev0 suffix from the version line in pyproject.toml.
2. `make test` to ensure tests pass.
3. `git commit -a -m 'Release version x.y.z'` at that release version.
4. `git tag x.y.z` at that release version.
5. `make dist` to build the bundles into the `dist/` folder.
6. `tar tf dist/*.tar.gz` to quickly inspect that all looks well.
7. `mamba deactivate; mamba activate dist` to switch to my twine env.
8. `twine upload dist/*` to upload the results to PyPI.
9. Bump the patch digit and append .dev0 suffix to the version line.
10. `git commit -a -m 'Bump to next development cycle'`.
11. `git push; git push origin x.y.z` to push main and the tag.

It looks like a lot when written out that way, but it's straightforward.