Drop support for Python 3.7
There is not a build of Python 3.7 available on macOS arm64
architecture, which it seems GitHub's macos-latest VMs are now using
instead of x64. And Python 3.7 is end-of-life anyway, so we bump the
minimum to 3.8.

This commit also update the upper-end CI tests from 3.10 to 3.12,
the newest stable release as of this writing.